(feat) Allow server to itself be called from the event loop
This is so (hopefully) will be able to test `run_application`. This does put a
bit more code outside of `run_application`, but as long as there is some higher
level testing, this should be fine.

There is less graceful shutdown on iterrupt here. This should be thought about.
However, since the system has to handle ungraceful shutdown _anyway_, say, if
the surrounding container just dies, then spending time keeping graceful
shutdown may not be worthwile. This isn't user-facing, so reducing chance of a
user seeing an error isn't a concern.