(feat) Simple polling
At this point it just prints the result of the poll. Also '127.0.0.1' is used
rather than 'localhost', since CircleCI seems to use '::1' for 'localhost', but
has issues listening on IPv6.

Tasks are all cancelled to make sure polling doesn't continue beyond the end of
each test (errors can be intermittantly triggered if this wasn't done).