(feat) Support multiple feed seed endpoints
There will be the same number of parallel requests to elastic search as there
are feed endpoints. This may be changed later, or might not, since there isn't
a plan for a high number of these endpoints (~6 or so).

Note that the order requests are received by the test Elasticsearch server is
not guarenteed. Even when testing locally, they would sometimes arrive in
different orders.