(build) Enforce a limit on number of arguments
The default pylint limit is 5, but every possibility of wrangling the arguments
of es_bulk_auth_headers just seems to make things less clear or safe: wrapping
in intermediate data structures ends up with more to reason about, not less.

I can't see anything in PEP8 about too many arguments, so it looks like it's
pylint-specific. And the codebase for pylint itself sets this to 10

```
https://github.com/PyCQA/pylint/blob/c21c731451e09c68f29d670bd286b23638269116/pylintrc#L302
```

The following leaps to mind:

> Oh, I have tasted my own medicine, and it is bitter!