(feat) IP Whitelist
Only uses the X-Forwarded-For header, since this is how the client IP is
identified in PaaS, which is where this is running, and a fair number of other
cloud providers in case it ever runs somewhere else.

Deliberately _not_ using a more complex mechanism of determining the client, to
minimise attack surface area. There is _one_ way of determining the source IP,
and this needs to be un-spoofable.

Local development is still possible by manually setting headers.