(feat) Constant Server header that doesn't expose versions
It's an extremely minor security issue (especially for a project that's open
source), but the ITHC does pick it up.

For some reason aiohttp makes it _really_ hard to just _not_ send the header.
The only was I could come up with would be to monkey-patch the library, which I
think would risk being too brittle (and potentially, insecure).