(feat) Ensure index and mapping after any exceptions
After the first iteration, the ensure_* functions will do nothing, so the
commit is _almost_ a refactor...

... but not quite, since some errors on index creation that would have bubbled
up will be swallowed (hence the changes in the tests). So, this could be argued
to be a regression of error handling. However, since the plan is to just create
a new index with mappings, on each iteration, then throwing an error on
creaation is less likely, since there will be no conflicting index to cause
this. For now, I think it's acceptable to do this (especially as this isn't
live!)

Reporting on such "ongoing" errors, post load of the application, is an
existing TODO.