(feat) No pause for activity stream feed
They're all our own endpoints, that aren't rate limited. By nature of how the
stream works, we only have max 1 concurrent request, so it's not a DoS attack:
it's more like 1 extra concurrent user, which the services _should_ really be
able to handle.

The test change is due to the fact killing the test ES server early means that
exceptions are attempted to be sent to Sentry, which aren't sent by the end of
its close, which cause some of its internal assertions to fail, and fail the
test. I'm not convinced that the ayncio raven client is that good, but too much
of a project right now to write our own.