(feat) Reduce number of (primary) shards from 5 to 4
There are 2 m4.large instances in the cluster, and each has 2 VCPUs. Making it
so each primary shard has ~exclusive use of a VCPU, to avoid overallocation.

Re-indexing is built into the algorithm, so this isn't in any way final. Can
just tweak and re-deploy if needs be.

The test change was due to the fact that they were flaky wrt to the order of
documents returned. This I think dependended on how they were split between
shards.