(feat) Use own http for higher speed and lower memory: non chunked for now
The structure is probably a bit more complex than is needed for right now, but
should hopefully fit into the framework needed for chunked encoding in the next
commit.

Note, the main complication is that we can never be sure how many bytes we can
get out of the socket. We may just get one at a time.