DET-177: add a raw keyword field for `object.name`
We want to sort activities by the `object.name` field. In order to do
this we need it to be a keyword, but we don't necessarily want to lose
the text type field because that will be useful for search. This commit
adds a `raw` field to the `name` that is of type keyword.

A search with the opensearch DSL below should be now be possible:

```
curl -XGET localhost:9200/_search \
-H "content-type: application/json" \
-d '{
    "query": {
      "match": {
        "type": "Create"
      }
    },
    "sort": {
        "object.name.raw": {
            "order": "asc",
            "unmapped_type": "string"
            }
        }
    }'
```