DET-232: Add standardised name fields to Aventri Attendees
We would like to be able to sort attendees for Aventri events by
first/last names.

There are existing fields on `dit:aventri:Attendees` objects with fields
`dit:aventri:lastName` and `dit:aventri:firstName`, but as we want to
avoid mapping feed-specific fields when they could be generic, we're creating
new fields called `dit:firstName` and `dit:lastName`.

This commit adds the new fields to the attendees.