feat: decrease number of shards per feed to 1 (plus replica)
This decreases the number of primary shards per feed to 1.

In production the largest shard is 1.1GB. According to
https://opensearch.org/blog/optimize-opensearch-index-shard-size/ shards should
be between 10GB and 50GB depending on load type - so I think we're safe to
increase.

The reason for this isn't just to align with the guidance: we're getting a lot
of 429s on searches. Each search is actually a "search on each shard" and so
each shard contributes, we think, to a some limit being reached.