fix: force empty content-type header for GET requests
Some services expect no content-type header for GET requests (or empty string),
and some completely don't allow a content-type header of the empty string.

So to force no content-type header, we pass None as the body of the request.
Otherwise recent versions of aiohttp pass application/octet-stream, which only
a few services allow.